+++ NEWS.md
@@ -1,4 +1,8 @@
 # News
+## 0.1.2
+* Function PF_lm.R has changed outputl, returns a list of three elements
+* Changes in the examples section of PF_lm.R documentation
+* Better performance achieved with the new function PF_lm_ss.R which uses method of simple sampling as resampling method 
 
 ## 0.1.1
 * Changes in the documentation